+++ engine/Engine.h
@@ -16,7 +16,7 @@ class Engine {
 private:
     std::string _name;
 
-    double _triPerSec = 0;
+    int _triPerSec = 0;
     bool _updateWorld = true;
     bool _debugText = Consts::SHOW_DEBUG_INFO;
     bool _useOpenGL = Consts::USE_OPEN_GL;
